requirements.txt
einops
ftfy
h5py
imageio
pytorch-lightning==1.5.4
regex
scikit-video
tqdm
fairscale
lightning
diffusers
accelerate
lpips
tensorflow
numpy==1.26.2
scipy
opencv-python
timm
decord
transformers
beartype
imageio[ffmpeg]
imageio[pyav]
scikit-learn
wandb
magvit2-pytorch
transformer-engine
torchmetrics==1.4.0.post0
torchmetrics[image]
pytorch-msssim
deepspeed==0.16.2
SentencePiece
torchshow